Boyscouting. Remove imports to UML package, docstrings
